+++ src/window_main/settings.js
@@ -492,7 +492,6 @@ function appendDataSection(section) {
 function appendOverlay(section) {
   section.appendChild(createDiv(["settings_title"], "Overlays"));
 
-  const displayControls = createDiv(["settings_row"]);
   // Toggle Edit Mode Button
   const editModeButton = createDiv(
     ["button_simple"],
@@ -502,32 +501,32 @@ function appendOverlay(section) {
   editModeButton.addEventListener("click", function() {
     ipcSend("toggle_edit_mode");
   });
-  displayControls.appendChild(editModeButton);
-  // Set Overlay Display Screen
-  const overlayDisplay = pd.settings.overlay_display
-    ? pd.settings.overlay_display
-    : remote.screen.getPrimaryDisplay().id;
-  const label = createLabel(["but_container_label"], "Overlay Display:");
-  label.style.marginTop = "auto";
-  const displaySelect = createSelect(
-    label,
-    remote.screen.getAllDisplays().map(display => display.id),
-    overlayDisplay,
-    filter => ipcSend("save_user_settings", { overlay_display: filter }),
-    "overlay_display",
-    filter => {
-      const displayNumber = remote.screen
-        .getAllDisplays()
-        .findIndex(d => d.id == filter);
-      const primary = filter == remote.screen.getPrimaryDisplay().id;
+  section.appendChild(editModeButton);
 
-      return `Display ${displayNumber} ${primary ? "(primary)" : ""}`;
-    }
+  // Copy pasta with bolognesa!
+  const pickerLabel = createLabel(
+    ["but_container_label"],
+    "<span style='margin-right: 32px;'>Background color <i>(0,0,0,0 to use default background)</i>:</span>"
   );
-  displaySelect.style.width = "180px";
-  displaySelect.style.marginLeft = "32px";
-  displayControls.appendChild(label);
-  section.appendChild(displayControls);
+  const colorPick = createInput(["color_picker"], "", {
+    id: "flat",
+    type: "text",
+    value: "Example Content"
+  });
+  colorPick.style.backgroundColor = pd.settings.overlay_back_color;
+  colorPick.addEventListener("click", function(e) {
+    e.stopPropagation();
+    showColorpicker(
+      pd.settings.overlay_back_color,
+      color => (colorPick.style.backgroundColor = color.rgbaString),
+      color =>
+        ipcSend("save_user_settings", { overlay_back_color: color.rgbaString }),
+      () => (colorPick.style.backgroundColor = pd.settings.overlay_back_color),
+      { alpha: true }
+    );
+  });
+  pickerLabel.appendChild(colorPick);
+  section.appendChild(pickerLabel);
 
   const sliderScale = createDiv(["slidecontainer_settings"]);
   const sliderScaleLabel = createLabel(